Documents/Mindset/local_deployment/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Data storage and processing
duckdb>=0.3.4
pyarrow>=6.0.0
dask[complete]>=2021.10.0
fastparquet>=0.8.0

# NLP and ML
transformers>=4.12.0
torch>=1.10.0
spacy>=3.0.0
nltk>=3.6.0
gensim>=4.1.2
xgboost>=1.5.0
lightgbm>=3.3.0
shap>=0.40.0

# API and web
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=1.8.0
requests>=2.26.0
python-dotenv>=0.19.0
newsapi-python>=0.2.6

# Testing and development
pytest>=6.2.0
black>=21.10b0
flake8>=4.0.0
mypy>=0.910
jupyter>=1.0.0
notebook>=6.4.0